pyproject.toml
[tool.poetry]
name = "birthdays"
version = "0.1.0"
description = ""
authors = ["Mohammed Aminu <aminuolawalekan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
celery = "^5.0.5"
django-graphql-jwt = "^0.3.1"
redis = "^3.5.3"
django-environ = "^0.4.5"
django-cors-headers = "^3.6.0"
psycopg2 = "^2.8.6"
django-templated-email = "^2.3.0"
PyJWT = "v1.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
